+++ src/dataFormat/field.ts
@@ -1,6 +1,12 @@
 import { camelCaseToSnakeCase, quote } from '../helpers/string'
 import { Field, FieldOptions, FieldsOptions } from './types/field'
 
+/**
+ * For creating a reusable group of fields that can be used across Data Formats.
+ *
+ * This is useful when multiple of your Data Formats share some common fields, like
+ * "dateCreated", "dataDeleted", "name", and so on.
+ */
 export const createCommonFields = <T extends FieldsOptions>(v: T): T => v
 
 export const createField = (fieldName: string, fieldOptions: FieldOptions): Field => {
