+++ src/helpers/types.ts
@@ -133,3 +133,5 @@ export type ArrayTernary<T, TIsArray extends boolean> = TIsArray extends false ?
 type IfAny<T, Y, N> = 0 extends (1 & T) ? Y : N
 
 export type IsAny<T> = IfAny<T, true, false>
+
+export type StringKeysOf<T> = Extract<keyof T, string>
